Backend: fix validation of requests, fix projection, fix AuctionEventSortServiceImpl, change logic for betting system + STOMP
FrontEnd: Fix UploadFiles, add styles for Header and Footer,

+++ Backend/src/main/java/com/auction/projection/AuctionEventSortProjection.java
@@ -1,9 +1,8 @@
 package com.auction.projection;
 
-import java.math.BigInteger;
 
 public interface AuctionEventSortProjection {
-  BigInteger getCount();
+  Long getCount();
 
-  BigInteger getAuctionId();
+  Long getAuctionId();
 }
